Support syntax highlighting
Also fix an issue where by it was preventing syntax highlighting from
working.

Also fixes using the preference as the rest api does not correctly
use the correct named type variables so we need to convert the
named variables into the supported ones codemirror side.

Also adds missing themes and addons and mode files.

Also reorganise the modes so that it's alphabetical.

See [1]

[1] https://github.com/GerritCodeReview/gerrit/blob/6c1a52b8698d1ad392c590e7b515f2b5668e73f4/gerrit-gwtui/src/main/java/com/google/gerrit/client/editor/EditScreen.java#L555

Change-Id: Icc222cde89eff7a0992fbcaaf8e7bac560347c18