Remove CodeMirror modes that are not supported at Google
If there is a desperate need for any of those modes, please reach out to
me. We are not opposed to adding a mode here and there, but it causes
work for Google maintainers. The unsupported modes don't seem to be
important enough to justify this work at the moment.

Change-Id: I9f47d6d3ba0383acabdd9d64825554c29fcee8c8