Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> do we have the stable-3.2 branch now ? if so, we may can start removing them for master

We don't, but submitting this will do no harm.  The submodule subscription will only kick in for the gerrit master branch. In other words, it won't get into gerrit stable-3.2.

If we want to do 3.2 specific fixes on this plugin later, we can create a stable-3.2 branch then.

Patch-set: 3
CC: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>