Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > Should a 2.16/3.0/3.1 branch be created before merging?
> > 
> > As I replied above, we don't need it yet.  We can create it when/if we need it.
> 
> I think we don't have any stable branches yet for this plugin, and with this change and polymer 3 changes, master will not support any version before 3.1, I think its a good time to create one now, wdyt ?

We only need to create the stable branches if/when we need to backport fixes for use in earlier versions of gerrit.  For now, we can keep the master branch only; changes submitted on master will only get included in gerrit's master due to how the submodule subscription works.

If we want to include the polymer 3 changes in 3.2, we can manually update the revision on gerrit's stable-3.2 branch.

This is how we have been doing it for all the other core plugins.

Patch-set: 5