Bazel: Don't use assets parameter in polygerrit_plugin rule
All assets parameter does is add its content to the output. We don't
really need it, as we add the assets into static directory of plugin
artifact in genrule anyway.

Given that this plugin is the only caller site that is using assets
parameter, we can kill it now in gerrit core and simplify
polygerrit_plugin rule.

Change-Id: I54d4a910631c9410883b7fd64ba0d88638f9d579