Improve performance
We stop setting viewportMargin and instead use the default value.

We also add support for the inbuilt search support within codemirror.
This is because with the removal of viewportMargin you cannot use
the browsers search to search for the code.

In my view i think this is a good trade off as you'll still be able to
search using the codemirror searching support whilst improving performance for large documents.

Change-Id: I84f12168c4784cf40ca102f025b7805b44bc9d9a