Fix mobile support
I'm not entirely sure if there is any unintended side effects.
I could only test in a simulator.

On iPhones doing any type of selection would break. Setting
the inputStyle to use textarea fixed the issue for me in the simulator.

On desktop inputStyle defaults to textarea but on mobile its
contenteditable.

Change-Id: I7f47be4b8c5029abafc830b987a8bf1719a2ea95