Display cursor position in the bottom
The line number the cursor is on and column number will be displayed.

The column number is more important as it means users don't
have to copy the line into a text editor to find the column number.

We did this in GWTUI [1] but we never did this for the new UI.

[1] If9090a893631077349a6ccf4f065f580dc778b08

Change-Id: Ibcac9d5e4283c8c7bd642420627d44eaac4d4f44